requirements.txt
torch
numpy
pandas
matplotlib
seaborn
scikit-learn
pycaret
tensorflow
ucimlrepo
Pillow